Crate webrender_traits

Modules

api

channel

display_list

units
A collection of coordinate spaces and their corresponding Point, Size and Rect types.

Structs

AuxiliaryLists

AuxiliaryListsDescriptor
Describes the memory layout of the auxiliary lists.

BorderDisplayItem

BorderRadius

BorderSide

BoxShadowDisplayItem

BuiltDisplayList
A display list.

BuiltDisplayListDescriptor
Describes the memory layout of a display list.

ClipRegion

ColorF

ColorU

ComplexClipRegion

DevicePixel
Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

DisplayItem

DisplayListBuilder

DynamicProperties
When using generate_frame(), a list of PropertyValue structures can optionally be supplied to provide the current value of any animated properties.

Epoch

ExternalEvent
An opaque pointer-sized value.

ExternalImageId
An arbitrary identifier for an external image provided by the application. It must be a unique identifier for each external image.

FontKey

GlyphDimensions

GlyphInstance

GlyphKey

GlyphOptions

GradientDisplayItem

GradientStop

IdNamespace

IframeDisplayItem

ImageDescriptor

ImageDisplayItem

ImageKey

ImageMask

ItemRange

LayerPixel
Geometry in a layer's local coordinate space (logical pixels).

NativeFontHandle
Native fonts are not used on Linux; all fonts are raw.

PipelineId

PropertyBindingId

PropertyBindingKey
A unique key that is used for connecting animated property values to bindings in the display list.

PropertyValue
The current value of an animated property. This is supplied by the calling code.

PushScrollLayerItem

PushStackingContextDisplayItem

RadialGradientDisplayItem

RectangleDisplayItem

RenderApi

RenderApiSender

ResourceId

ScrollLayerId

ScrollLayerPixel
Geometry in a layer's scrollable parent coordinate space (logical pixels).

ScrollLayerState

ServoScrollRootId

StackingContext

SubpixelPoint

TextDisplayItem

WebGLBufferId

WebGLContextId

WebGLDisplayItem

WebGLFramebufferId

WebGLProgramId

WebGLRenderbufferId

WebGLShaderId

WebGLTextureId

WorldPixel
Geometry in the document's coordinate space (logical pixels).

YuvImageDisplayItem

Enums

ApiMsg

BorderStyle

BoxShadowClipMode

DisplayListMode

ExtendMode

FilterOp

FontRenderMode

FragmentType

ImageData

ImageFormat

ImageRendering

MixBlendMode

PropertyBinding
A binding property can either be a specific value (the normal, non-animated case) or point to a binding location to fetch the current value from.

RendererKind

ScrollEventPhase

ScrollLayerInfo

ScrollLocation

ScrollPolicy

SpecificDisplayItem

SubpixelOffset

VRCompositorCommand

WebGLCommand

WebGLError

WebGLFramebufferBindingRequest

WebGLParameter

WebGLShaderParameter

YuvColorSpace

Constants

_IMPL_DESERIALIZE_FOR_ApiMsg

_IMPL_DESERIALIZE_FOR_AuxiliaryLists

_IMPL_DESERIALIZE_FOR_AuxiliaryListsDescriptor

_IMPL_DESERIALIZE_FOR_BorderDisplayItem

_IMPL_DESERIALIZE_FOR_BorderRadius

_IMPL_DESERIALIZE_FOR_BorderSide

_IMPL_DESERIALIZE_FOR_BorderStyle

_IMPL_DESERIALIZE_FOR_BoxShadowClipMode

_IMPL_DESERIALIZE_FOR_BoxShadowDisplayItem

_IMPL_DESERIALIZE_FOR_BuiltDisplayList

_IMPL_DESERIALIZE_FOR_BuiltDisplayListDescriptor

_IMPL_DESERIALIZE_FOR_ClipRegion

_IMPL_DESERIALIZE_FOR_ColorF

_IMPL_DESERIALIZE_FOR_ColorU

_IMPL_DESERIALIZE_FOR_ComplexClipRegion

_IMPL_DESERIALIZE_FOR_DisplayItem

_IMPL_DESERIALIZE_FOR_DisplayListMode

_IMPL_DESERIALIZE_FOR_DynamicProperties

_IMPL_DESERIALIZE_FOR_Epoch

_IMPL_DESERIALIZE_FOR_ExtendMode

_IMPL_DESERIALIZE_FOR_ExternalEvent

_IMPL_DESERIALIZE_FOR_ExternalImageId

_IMPL_DESERIALIZE_FOR_FilterOp

_IMPL_DESERIALIZE_FOR_FontKey

_IMPL_DESERIALIZE_FOR_FontRenderMode

_IMPL_DESERIALIZE_FOR_FragmentType

_IMPL_DESERIALIZE_FOR_GlyphDimensions

_IMPL_DESERIALIZE_FOR_GlyphInstance

_IMPL_DESERIALIZE_FOR_GlyphKey

_IMPL_DESERIALIZE_FOR_GlyphOptions

_IMPL_DESERIALIZE_FOR_GradientDisplayItem

_IMPL_DESERIALIZE_FOR_GradientStop

_IMPL_DESERIALIZE_FOR_IdNamespace

_IMPL_DESERIALIZE_FOR_IframeDisplayItem

_IMPL_DESERIALIZE_FOR_ImageData

_IMPL_DESERIALIZE_FOR_ImageDescriptor

_IMPL_DESERIALIZE_FOR_ImageDisplayItem

_IMPL_DESERIALIZE_FOR_ImageFormat

_IMPL_DESERIALIZE_FOR_ImageKey

_IMPL_DESERIALIZE_FOR_ImageMask

_IMPL_DESERIALIZE_FOR_ImageRendering

_IMPL_DESERIALIZE_FOR_ItemRange

_IMPL_DESERIALIZE_FOR_MixBlendMode

_IMPL_DESERIALIZE_FOR_NativeFontHandle

_IMPL_DESERIALIZE_FOR_PipelineId

_IMPL_DESERIALIZE_FOR_PropertyBinding

_IMPL_DESERIALIZE_FOR_PropertyBindingId

_IMPL_DESERIALIZE_FOR_PropertyBindingKey

_IMPL_DESERIALIZE_FOR_PropertyValue

_IMPL_DESERIALIZE_FOR_PushScrollLayerItem

_IMPL_DESERIALIZE_FOR_PushStackingContextDisplayItem

_IMPL_DESERIALIZE_FOR_RadialGradientDisplayItem

_IMPL_DESERIALIZE_FOR_RectangleDisplayItem

_IMPL_DESERIALIZE_FOR_RenderApiSender

_IMPL_DESERIALIZE_FOR_ResourceId

_IMPL_DESERIALIZE_FOR_ScrollEventPhase

_IMPL_DESERIALIZE_FOR_ScrollLayerId

_IMPL_DESERIALIZE_FOR_ScrollLayerInfo

_IMPL_DESERIALIZE_FOR_ScrollLayerState

_IMPL_DESERIALIZE_FOR_ScrollLocation

_IMPL_DESERIALIZE_FOR_ScrollPolicy

_IMPL_DESERIALIZE_FOR_ServoScrollRootId

_IMPL_DESERIALIZE_FOR_SpecificDisplayItem

_IMPL_DESERIALIZE_FOR_StackingContext

_IMPL_DESERIALIZE_FOR_SubpixelOffset

_IMPL_DESERIALIZE_FOR_SubpixelPoint

_IMPL_DESERIALIZE_FOR_TextDisplayItem

_IMPL_DESERIALIZE_FOR_VRCompositorCommand

_IMPL_DESERIALIZE_FOR_WebGLCommand

_IMPL_DESERIALIZE_FOR_WebGLContextId

_IMPL_DESERIALIZE_FOR_WebGLDisplayItem

_IMPL_DESERIALIZE_FOR_WebGLError

_IMPL_DESERIALIZE_FOR_WebGLFramebufferBindingRequest

_IMPL_DESERIALIZE_FOR_WebGLParameter

_IMPL_DESERIALIZE_FOR_WebGLShaderParameter

_IMPL_DESERIALIZE_FOR_YuvColorSpace

_IMPL_DESERIALIZE_FOR_YuvImageDisplayItem

_IMPL_SERIALIZE_FOR_ApiMsg

_IMPL_SERIALIZE_FOR_AuxiliaryLists

_IMPL_SERIALIZE_FOR_AuxiliaryListsDescriptor

_IMPL_SERIALIZE_FOR_BorderDisplayItem

_IMPL_SERIALIZE_FOR_BorderRadius

_IMPL_SERIALIZE_FOR_BorderSide

_IMPL_SERIALIZE_FOR_BorderStyle

_IMPL_SERIALIZE_FOR_BoxShadowClipMode

_IMPL_SERIALIZE_FOR_BoxShadowDisplayItem

_IMPL_SERIALIZE_FOR_BuiltDisplayList

_IMPL_SERIALIZE_FOR_BuiltDisplayListDescriptor

_IMPL_SERIALIZE_FOR_ClipRegion

_IMPL_SERIALIZE_FOR_ColorF

_IMPL_SERIALIZE_FOR_ColorU

_IMPL_SERIALIZE_FOR_ComplexClipRegion

_IMPL_SERIALIZE_FOR_DisplayItem

_IMPL_SERIALIZE_FOR_DisplayListMode

_IMPL_SERIALIZE_FOR_DynamicProperties

_IMPL_SERIALIZE_FOR_Epoch

_IMPL_SERIALIZE_FOR_ExtendMode

_IMPL_SERIALIZE_FOR_ExternalEvent

_IMPL_SERIALIZE_FOR_ExternalImageId

_IMPL_SERIALIZE_FOR_FilterOp

_IMPL_SERIALIZE_FOR_FontKey

_IMPL_SERIALIZE_FOR_FontRenderMode

_IMPL_SERIALIZE_FOR_FragmentType

_IMPL_SERIALIZE_FOR_GlyphDimensions

_IMPL_SERIALIZE_FOR_GlyphInstance

_IMPL_SERIALIZE_FOR_GlyphKey

_IMPL_SERIALIZE_FOR_GlyphOptions

_IMPL_SERIALIZE_FOR_GradientDisplayItem

_IMPL_SERIALIZE_FOR_GradientStop

_IMPL_SERIALIZE_FOR_IdNamespace

_IMPL_SERIALIZE_FOR_IframeDisplayItem

_IMPL_SERIALIZE_FOR_ImageData

_IMPL_SERIALIZE_FOR_ImageDescriptor

_IMPL_SERIALIZE_FOR_ImageDisplayItem

_IMPL_SERIALIZE_FOR_ImageFormat

_IMPL_SERIALIZE_FOR_ImageKey

_IMPL_SERIALIZE_FOR_ImageMask

_IMPL_SERIALIZE_FOR_ImageRendering

_IMPL_SERIALIZE_FOR_ItemRange

_IMPL_SERIALIZE_FOR_MixBlendMode

_IMPL_SERIALIZE_FOR_NativeFontHandle

_IMPL_SERIALIZE_FOR_PipelineId

_IMPL_SERIALIZE_FOR_PropertyBinding

_IMPL_SERIALIZE_FOR_PropertyBindingId

_IMPL_SERIALIZE_FOR_PropertyBindingKey

_IMPL_SERIALIZE_FOR_PropertyValue

_IMPL_SERIALIZE_FOR_PushScrollLayerItem

_IMPL_SERIALIZE_FOR_PushStackingContextDisplayItem

_IMPL_SERIALIZE_FOR_RadialGradientDisplayItem

_IMPL_SERIALIZE_FOR_RectangleDisplayItem

_IMPL_SERIALIZE_FOR_RenderApiSender

_IMPL_SERIALIZE_FOR_ResourceId

_IMPL_SERIALIZE_FOR_ScrollEventPhase

_IMPL_SERIALIZE_FOR_ScrollLayerId

_IMPL_SERIALIZE_FOR_ScrollLayerInfo

_IMPL_SERIALIZE_FOR_ScrollLayerState

_IMPL_SERIALIZE_FOR_ScrollLocation

_IMPL_SERIALIZE_FOR_ScrollPolicy

_IMPL_SERIALIZE_FOR_ServoScrollRootId

_IMPL_SERIALIZE_FOR_SpecificDisplayItem

_IMPL_SERIALIZE_FOR_StackingContext

_IMPL_SERIALIZE_FOR_SubpixelOffset

_IMPL_SERIALIZE_FOR_SubpixelPoint

_IMPL_SERIALIZE_FOR_TextDisplayItem

_IMPL_SERIALIZE_FOR_VRCompositorCommand

_IMPL_SERIALIZE_FOR_WebGLCommand

_IMPL_SERIALIZE_FOR_WebGLContextId

_IMPL_SERIALIZE_FOR_WebGLDisplayItem

_IMPL_SERIALIZE_FOR_WebGLError

_IMPL_SERIALIZE_FOR_WebGLFramebufferBindingRequest

_IMPL_SERIALIZE_FOR_WebGLParameter

_IMPL_SERIALIZE_FOR_WebGLShaderParameter

_IMPL_SERIALIZE_FOR_YuvColorSpace

_IMPL_SERIALIZE_FOR_YuvImageDisplayItem

Traits

RenderDispatcher

RenderNotifier

VRCompositorHandler

Functions

as_scroll_parent_rect

device_length

Type Definitions

DeviceIntLength

DeviceIntPoint

DeviceIntRect

DeviceIntSize

DevicePoint

DeviceRect

DeviceSize

DeviceUintPoint

DeviceUintRect

DeviceUintSize

LayerPoint

LayerPoint4D

LayerRect

LayerSize

LayerToScrollTransform

LayerToWorldTransform

LayerTransform

LayoutPixel
Geometry in a stacking context's local coordinate space (logical pixels).

LayoutPoint

LayoutRect

LayoutSize

LayoutTransform

ScrollLayerPoint

ScrollLayerRect

ScrollLayerSize

ScrollToLayerTransform

ScrollToWorldTransform

VRCompositorId

WebGLResult

WorldPoint

WorldPoint4D

WorldRect

WorldSize

WorldToLayerTransform