Module webrender_traits::api

Structs

DynamicProperties
When using generate_frame(), a list of PropertyValue structures can optionally be supplied to provide the current value of any animated properties.

Epoch

ExternalEvent
An opaque pointer-sized value.

IdNamespace

PipelineId

PropertyBindingId

PropertyBindingKey
A unique key that is used for connecting animated property values to bindings in the display list.

PropertyValue
The current value of an animated property. This is supplied by the calling code.

RenderApi

RenderApiSender

ResourceId

ScrollLayerState

ZoomFactor
Represents a zoom factor.

Enums

ApiMsg

PropertyBinding
A binding property can either be a specific value (the normal, non-animated case) or point to a binding location to fetch the current value from.

ScrollEventPhase

ScrollLocation

VRCompositorCommand

Constants

_IMPL_DESERIALIZE_FOR_ApiMsg

_IMPL_DESERIALIZE_FOR_DynamicProperties

_IMPL_DESERIALIZE_FOR_Epoch

_IMPL_DESERIALIZE_FOR_ExternalEvent

_IMPL_DESERIALIZE_FOR_IdNamespace

_IMPL_DESERIALIZE_FOR_PipelineId

_IMPL_DESERIALIZE_FOR_PropertyBinding

_IMPL_DESERIALIZE_FOR_PropertyBindingId

_IMPL_DESERIALIZE_FOR_PropertyBindingKey

_IMPL_DESERIALIZE_FOR_PropertyValue

_IMPL_DESERIALIZE_FOR_RenderApiSender

_IMPL_DESERIALIZE_FOR_ResourceId

_IMPL_DESERIALIZE_FOR_ScrollEventPhase

_IMPL_DESERIALIZE_FOR_ScrollLayerState

_IMPL_DESERIALIZE_FOR_ScrollLocation

_IMPL_DESERIALIZE_FOR_VRCompositorCommand

_IMPL_DESERIALIZE_FOR_ZoomFactor

_IMPL_SERIALIZE_FOR_ApiMsg

_IMPL_SERIALIZE_FOR_DynamicProperties

_IMPL_SERIALIZE_FOR_Epoch

_IMPL_SERIALIZE_FOR_ExternalEvent

_IMPL_SERIALIZE_FOR_IdNamespace

_IMPL_SERIALIZE_FOR_PipelineId

_IMPL_SERIALIZE_FOR_PropertyBinding

_IMPL_SERIALIZE_FOR_PropertyBindingId

_IMPL_SERIALIZE_FOR_PropertyBindingKey

_IMPL_SERIALIZE_FOR_PropertyValue

_IMPL_SERIALIZE_FOR_RenderApiSender

_IMPL_SERIALIZE_FOR_ResourceId

_IMPL_SERIALIZE_FOR_ScrollEventPhase

_IMPL_SERIALIZE_FOR_ScrollLayerState

_IMPL_SERIALIZE_FOR_ScrollLocation

_IMPL_SERIALIZE_FOR_VRCompositorCommand

_IMPL_SERIALIZE_FOR_ZoomFactor

Traits

RenderDispatcher

RenderNotifier

VRCompositorHandler

Type Definitions

TileSize

VRCompositorId