Enum webrender_traits::api::ApiMsg

pub enum ApiMsg {
    AddRawFont(FontKey, Vec<u8>),
    AddNativeFont(FontKey, NativeFontHandle),
    AddImage(ImageKey, u32, u32, ImageFormat, Vec<u8>),
    UpdateImage(ImageKey, u32, u32, ImageFormat, Vec<u8>),
    CloneApi(IpcSender<IdNamespace>),
    SetRootStackingContext(StackingContextId, ColorF, Epoch, PipelineId, Size2D<f32>, Vec<(StackingContextId, StackingContext)>, Vec<(DisplayListId, BuiltDisplayListDescriptor)>, AuxiliaryListsDescriptor),
    SetRootPipeline(PipelineId),
    Scroll(Point2D<f32>, Point2D<f32>, ScrollEventPhase),
    TickScrollingBounce,
    TranslatePointToLayerSpace(Point2D<f32>, IpcSender<(Point2D<f32>, PipelineId)>),
    RequestWebGLContext(Size2D<i32>, GLContextAttributes, IpcSender<Result<(WebGLContextId, GLLimits), String>>),
    WebGLCommand(WebGLContextId, WebGLCommand),
}

Variants

AddRawFont

AddNativeFont

AddImage

UpdateImage

CloneApi

SetRootStackingContext

Supplies a new frame to WebRender.

The first StackingContextId describes the root stacking context. The actual stacking contexts are supplied as the sixth parameter, while the display lists that make up those stacking contexts are supplied as the seventh parameter.

After receiving this message, WebRender will read the display lists, followed by the auxiliary lists, from the payload channel.

SetRootPipeline

Scroll

TickScrollingBounce

TranslatePointToLayerSpace

RequestWebGLContext

WebGLCommand