Module webrender_traits::display_list

Defines data structures which are consumed by the Compositor.

Structs

CompositorDisplayListInfo

A data structure which stores compositor-side information about display lists sent to the compositor.

HitTestInfo

Information that Servo keeps alongside WebRender display items in order to add more context to hit test results.

ScrollTree

A tree of spatial nodes, which mirrors the spatial nodes in the WebRender display list, except these are used to scrolling in the compositor so that new offsets can be sent to WebRender.

ScrollTreeNode

A node in a tree of scroll nodes. This may either be a scrollable node which responds to scroll events or a non-scrollable one.

ScrollTreeNodeId

An id for a ScrollTreeNode in the ScrollTree. This contains both the index to the node in the tree’s array of nodes as well as the corresponding SpatialId for the SpatialNode in the WebRender display list.

ScrollableNodeInfo

Data stored for nodes in the ScrollTree that actually scroll, as opposed to reference frames and sticky nodes which do not.

Enums

ScrollSensitivity

The scroll sensitivity of a scroll node ie whether it can be scrolled due to input event and script events or only script events.