Struct webrender_traits::display_list::AuxiliaryListsDescriptor

pub struct AuxiliaryListsDescriptor {
    gradient_stops_size: usize,
    complex_clip_regions_size: usize,
    filters_size: usize,
    glyph_instances_size: usize,
}

Describes the memory layout of the auxiliary lists.

Auxiliary lists consist of some number of gradient stops, complex clip regions, filters, and glyph instances, in that order.

Fields

gradient_stops_size

complex_clip_regions_size

filters_size

glyph_instances_size

Methods

impl AuxiliaryListsDescriptor

fn size(&self) -> usize