Enum webrender_traits::ApiMsg

pub enum ApiMsg {
    AddRawFont(FontKey, Vec<u8>),
    AddNativeFont(FontKey, NativeFontHandle),
    GetGlyphDimensions(Vec<GlyphKey>, MsgSender<Vec<Option<GlyphDimensions>>>),
    AddImage(ImageKey, u32, u32, Option<u32>, ImageFormat, ImageData),
    UpdateImage(ImageKey, u32, u32, ImageFormat, Vec<u8>),
    DeleteImage(ImageKey),
    CloneApi(MsgSender<IdNamespace>),
    Flush,
    SetRootDisplayList(ColorF, Epoch, PipelineId, Size2D<f32>, BuiltDisplayListDescriptor, AuxiliaryListsDescriptor),
    SetRootPipeline(PipelineId),
    Scroll(Point2D<f32>, Point2D<f32>, ScrollEventPhase),
    ScrollLayersWithScrollId(Point2D<f32>, PipelineId, ServoScrollRootId),
    TickScrollingBounce,
    TranslatePointToLayerSpace(Point2D<f32>, MsgSender<(Point2D<f32>, PipelineId)>),
    GetScrollLayerState(MsgSender<Vec<ScrollLayerState>>),
    RequestWebGLContext(Size2D<i32>, GLContextAttributes, MsgSender<Result<(WebGLContextId, GLLimits), String>>),
    ResizeWebGLContext(WebGLContextId, Size2D<i32>),
    WebGLCommand(WebGLContextId, WebGLCommand),
    GenerateFrame,
}

Variants
AddRawFont(FontKey, Vec<u8>)AddNativeFont(FontKey, NativeFontHandle)GetGlyphDimensions(Vec<GlyphKey>, MsgSender<Vec<Option<GlyphDimensions>>>)

Gets the glyph dimensions

AddImage(ImageKey, u32, u32, Option<u32>, ImageFormat, ImageData)

Adds an image from the resource cache.

UpdateImage(ImageKey, u32, u32, ImageFormat, Vec<u8>)

Updates the the resource cache with the new image data.

DeleteImage(ImageKey)

Drops an image from the resource cache.

CloneApi(MsgSender<IdNamespace>)FlushSetRootDisplayList(ColorF, Epoch, PipelineId, Size2D<f32>, BuiltDisplayListDescriptor, AuxiliaryListsDescriptor)

Supplies a new frame to WebRender.

After receiving this message, WebRender will read the display list, followed by the auxiliary lists, from the payload channel.

SetRootPipeline(PipelineId)Scroll(Point2D<f32>, Point2D<f32>, ScrollEventPhase)ScrollLayersWithScrollId(Point2D<f32>, PipelineId, ServoScrollRootId)TickScrollingBounceTranslatePointToLayerSpace(Point2D<f32>, MsgSender<(Point2D<f32>, PipelineId)>)GetScrollLayerState(MsgSender<Vec<ScrollLayerState>>)RequestWebGLContext(Size2D<i32>, GLContextAttributes, MsgSender<Result<(WebGLContextId, GLLimits), String>>)ResizeWebGLContext(WebGLContextId, Size2D<i32>)WebGLCommand(WebGLContextId, WebGLCommand)GenerateFrame