Module webrender_traits::image

Structs

BlobImageDescriptor

ExternalImageId
An arbitrary identifier for an external image provided by the application. It must be a unique identifier for each external image.

ImageDescriptor

ImageKey

RasterizedBlobImage

Enums

BlobImageError

ImageData

ImageFormat

Constants

_IMPL_DESERIALIZE_FOR_ExternalImageId

_IMPL_DESERIALIZE_FOR_ImageData

_IMPL_DESERIALIZE_FOR_ImageDescriptor

_IMPL_DESERIALIZE_FOR_ImageFormat

_IMPL_DESERIALIZE_FOR_ImageKey

_IMPL_SERIALIZE_FOR_ExternalImageId

_IMPL_SERIALIZE_FOR_ImageData

_IMPL_SERIALIZE_FOR_ImageDescriptor

_IMPL_SERIALIZE_FOR_ImageFormat

_IMPL_SERIALIZE_FOR_ImageKey

Traits

BlobImageRenderer

Type Definitions

BlobImageData

BlobImageResult