Struct webrender_traits::rendering_context::RenderingContext

pub struct RenderingContext(Rc<RenderingContextData>);

A Servo rendering context, which holds all of the information needed to render Servo’s layout, and bridges WebRender and surfman.

Tuple Fields
0: Rc<RenderingContextData>
Implementations

impl RenderingContext

pub fn create( connection: &Connection, adapter: &Adapter, surface_type: SurfaceType<NativeWidget>, ) -> Result<Self, Error>

pub fn create_surface( &self, surface_type: SurfaceType<NativeWidget>, ) -> Result<Surface, Error>

pub fn destroy_surface(&self, surface: Surface) -> Result<(), Error>

pub fn create_surface_texture( &self, surface: Surface, ) -> Result<SurfaceTexture, Error>

pub fn destroy_surface_texture( &self, surface_texture: SurfaceTexture, ) -> Result<Surface, Error>

pub fn make_gl_context_current(&self) -> Result<(), Error>

pub fn swap_chain(&self) -> Result<&SwapChain<Device>, Error>

pub fn resize(&self, size: Size2D<i32>) -> Result<(), Error>

pub fn present(&self) -> Result<(), Error>

pub fn with_front_buffer<F: FnOnce(&Device, Surface) -> Surface>(&self, f: F)

Invoke a closure with the surface associated with the current front buffer. This can be used to create a surfman::SurfaceTexture to blit elsewhere.

pub fn device(&self) -> Ref<'_, Device>

pub fn connection(&self) -> Connection

pub fn adapter(&self) -> Adapter

pub fn native_context(&self) -> NativeContext

pub fn native_device(&self) -> NativeDevice

pub fn context_attributes(&self) -> ContextAttributes

pub fn context_surface_info(&self) -> Result<Option<SurfaceInfo>, Error>

pub fn surface_info(&self, surface: &Surface) -> SurfaceInfo

pub fn surface_texture_object(&self, surface: &SurfaceTexture) -> u32

pub fn get_proc_address(&self, name: &str) -> *const c_void

pub fn unbind_native_surface_from_context(&self) -> Result<(), Error>

pub fn bind_native_surface_to_context( &self, native_widget: NativeWidget, ) -> Result<(), Error>