Struct webrender_traits::CrossProcessCompositorApi

pub struct CrossProcessCompositorApi(pub IpcSender<CrossProcessCompositorMessage>);

A mechanism to send messages from ScriptThread to the parent process’ WebRender instance.

Tuple Fields
0: IpcSender<CrossProcessCompositorMessage>
Implementations

impl CrossProcessCompositorApi

pub fn dummy() -> Self

Create a new CrossProcessCompositorApi struct that does not have a listener on the other end to use for unit testing.

pub fn sender(&self) -> &IpcSender<CrossProcessCompositorMessage>

Get the sender for this proxy.

pub fn send_initial_transaction(&self, pipeline: WebRenderPipelineId)

Inform WebRender of the existence of this pipeline.

pub fn send_scroll_node( &self, pipeline_id: WebRenderPipelineId, point: LayoutPoint, scroll_id: ExternalScrollId, )

Perform a scroll operation.

pub fn send_display_list( &self, display_list_info: CompositorDisplayListInfo, list: BuiltDisplayList, )

Inform WebRender of a new display list for the given pipeline.

pub fn hit_test( &self, pipeline: Option<WebRenderPipelineId>, point: DevicePoint, flags: HitTestFlags, ) -> Vec<CompositorHitTestResult>

Perform a hit test operation. Blocks until the operation is complete and and a result is available.

pub fn generate_image_key(&self) -> Option<ImageKey>

Create a new image key. Blocks until the key is available.

pub fn add_image( &self, key: ImageKey, descriptor: ImageDescriptor, data: SerializableImageData, )

pub fn update_images(&self, updates: Vec<ImageUpdate>)

Perform an image resource update operation.

pub fn remove_unused_font_resources( &self, keys: Vec<FontKey>, instance_keys: Vec<FontInstanceKey>, )

pub fn add_font_instance( &self, font_instance_key: FontInstanceKey, font_key: FontKey, size: f32, flags: FontInstanceFlags, )

pub fn add_font( &self, font_key: FontKey, data: Arc<IpcSharedMemory>, index: u32, )

pub fn add_system_font(&self, font_key: FontKey, handle: NativeFontHandle)

pub fn fetch_font_keys( &self, number_of_font_keys: usize, number_of_font_instance_keys: usize, ) -> (Vec<FontKey>, Vec<FontInstanceKey>)