Struct webrender_traits::RenderApi

pub struct RenderApi {
    pub api_sender: IpcSender<ApiMsg>,
    pub payload_sender: IpcBytesSender,
    pub id_namespace: IdNamespace,
    pub next_id: Cell<ResourceId>,
}

Fields

api_sender

payload_sender

id_namespace

next_id

Methods

impl RenderApi

fn clone_sender(&self) -> RenderApiSender

fn add_raw_font(&self, bytes: Vec<u8>) -> FontKey

fn add_native_font(&self, native_font_handle: NativeFontHandle) -> FontKey

fn alloc_image(&self) -> ImageKey

fn add_image(&self, width: u32, height: u32, format: ImageFormat, bytes: Vec<u8>) -> ImageKey

fn update_image(&self, key: ImageKey, width: u32, height: u32, format: ImageFormat, bytes: Vec<u8>)

fn set_root_pipeline(&self, pipeline_id: PipelineId)

fn set_root_stacking_context(&self, stacking_context_id: StackingContextId, background_color: ColorF, epoch: Epoch, pipeline_id: PipelineId, viewport_size: Size2D<f32>, stacking_contexts: Vec<(StackingContextId, StackingContext)>, display_lists: Vec<(DisplayListId, BuiltDisplayList)>, auxiliary_lists: AuxiliaryLists)

Supplies a new frame to WebRender.

Arguments: * stacking_context_id: The ID of the root stacking context. * background_color: The background color of this pipeline. * epoch: A monotonically increasing timestamp. * pipeline_id: The ID of the pipeline that is supplying this display list. * viewport_size: The size of the viewport for this frame. * stacking_contexts: Stacking contexts used in this frame. * display_lists: Display lists used in this frame. * auxiliary_lists: Various items that the display lists and stacking contexts reference.

fn scroll(&self, delta: Point2D<f32>, cursor: Point2D<f32>, phase: ScrollEventPhase)

fn tick_scrolling_bounce_animations(&self)

fn translate_point_to_layer_space(&self, point: &Point2D<f32>) -> (Point2D<f32>, PipelineId)

fn get_scroll_layer_state(&self) -> Vec<ScrollLayerState>

fn request_webgl_context(&self, size: &Size2D<i32>, attributes: GLContextAttributes) -> Result<(WebGLContextId, GLLimits), String>

fn send_webgl_command(&self, context_id: WebGLContextId, command: WebGLCommand)

fn next_stacking_context_id(&self) -> StackingContextId

fn next_display_list_id(&self) -> DisplayListId

fn next_unique_id(&self) -> (u32, u32)