Trait webrender_traits::WebrenderExternalImageApi

pub trait WebrenderExternalImageApi {
    // Required methods
    fn lock(&mut self, id: u64) -> (WebrenderImageSource<'_>, Size2D<i32>);
    fn unlock(&mut self, id: u64);
}

This trait is used as a bridge between the different GL clients in Servo that handles WebRender ExternalImages and the WebRender ExternalImageHandler API. This trait is used to notify lock/unlock messages and get the required info that WR needs.

Required Methods

fn lock(&mut self, id: u64) -> (WebrenderImageSource<'_>, Size2D<i32>)

fn unlock(&mut self, id: u64)