Module webrender_traits::units

A collection of coordinate spaces and their corresponding Point, Size and Rect types.

Physical pixels take into account the device pixel ratio and their dimensions tend to correspond to the allocated size of resources in memory, while logical pixels don't have the device pixel ratio applied which means they are agnostic to the usage of hidpi screens and the like.

The terms "layer" and "stacking context" can be used interchangeably in the context of coordinate systems.

Structs

DevicePixel
Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

LayerPixel
Geometry in a layer's local coordinate space (logical pixels).

ScrollLayerPixel
Geometry in a layer's scrollable parent coordinate space (logical pixels).

WorldPixel
Geometry in the document's coordinate space (logical pixels).

Functions

as_scroll_parent_rect

device_length

Type Definitions

DeviceIntLength

DeviceIntPoint

DeviceIntRect

DeviceIntSize

DevicePoint

DeviceRect

DeviceSize

DeviceUintPoint

DeviceUintRect

DeviceUintSize

LayerPoint

LayerPoint4D

LayerRect

LayerSize

LayerToScrollTransform

LayerToWorldTransform

LayerTransform

ScrollLayerPoint

ScrollLayerRect

ScrollLayerSize

ScrollToLayerTransform

ScrollToWorldTransform

WorldPoint

WorldPoint4D

WorldRect

WorldSize

WorldToLayerTransform