Crate webrender_traits

Modules

api

channel

color

display_item

display_list

font

image

units
A collection of coordinate spaces and their corresponding Point, Size and Rect types.

webgl

Structs

AuxiliaryLists

AuxiliaryListsBuilder

AuxiliaryListsDescriptor
Describes the memory layout of the auxiliary lists.

BlobImageDescriptor

BorderDisplayItem

BorderRadius

BorderSide

BorderWidths

BoxShadowDisplayItem

BuiltDisplayList
A display list.

BuiltDisplayListDescriptor
Describes the memory layout of a display list.

ClipRegion

ColorF

ColorU

ComplexClipRegion

DevicePixel
Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

DisplayItem

DisplayListBuilder

DynamicProperties
When using generate_frame(), a list of PropertyValue structures can optionally be supplied to provide the current value of any animated properties.

Epoch

ExternalEvent
An opaque pointer-sized value.

ExternalImageId
An arbitrary identifier for an external image provided by the application. It must be a unique identifier for each external image.

FontKey

GlyphDimensions

GlyphInstance

GlyphKey

GlyphOptions

Gradient

GradientBorder

GradientDisplayItem

GradientStop

IdNamespace

IframeDisplayItem

ImageBorder

ImageDescriptor

ImageDisplayItem

ImageKey

ImageMask

ItemRange

LayerPixel
Geometry in a layer's local coordinate space (logical pixels).

NativeFontHandle
Native fonts are not used on Linux; all fonts are raw.

NinePatchDescriptor

NormalBorder

PipelineId

PropertyBindingId

PropertyBindingKey
A unique key that is used for connecting animated property values to bindings in the display list.

PropertyValue
The current value of an animated property. This is supplied by the calling code.

PushScrollLayerItem

PushStackingContextDisplayItem

RadialGradient

RadialGradientBorder

RadialGradientDisplayItem

RasterizedBlobImage

RectangleDisplayItem

RenderApi

RenderApiSender

ResourceId

ScrollLayerId

ScrollLayerPixel
Geometry in a layer's scrollable parent coordinate space (logical pixels).

ScrollLayerState

ServoScrollRootId

StackingContext

SubpixelPoint

TextDisplayItem

Tiles
Offset in number of tiles.

WebGLBufferId

WebGLContextId

WebGLDisplayItem

WebGLFramebufferId

WebGLProgramId

WebGLRenderbufferId

WebGLShaderId

WebGLTextureId

WorldPixel
Geometry in the document's coordinate space (logical pixels).

YuvImageDisplayItem

ZoomFactor
Represents a zoom factor.

Enums

ApiMsg

BlobImageError

BorderDetails

BorderStyle

BoxShadowClipMode

ExtendMode

FilterOp

FontRenderMode

ImageData

ImageFormat

ImageRendering

MixBlendMode

PropertyBinding
A binding property can either be a specific value (the normal, non-animated case) or point to a binding location to fetch the current value from.

RepeatMode

ScrollEventPhase

ScrollLayerInfo

ScrollLocation

ScrollPolicy

SpecificDisplayItem

SubpixelOffset

VRCompositorCommand

WebGLCommand

WebGLError

WebGLFramebufferBindingRequest

WebGLParameter

WebGLShaderParameter

YuvColorSpace

Traits

BlobImageRenderer

RenderDispatcher

RenderNotifier

VRCompositorHandler

Functions

as_scroll_parent_rect

device_length

Type Definitions

BlobImageData

BlobImageResult

DeviceIntLength

DeviceIntPoint

DeviceIntRect

DeviceIntSize

DevicePoint

DeviceRect

DeviceSize

DeviceUintPoint

DeviceUintRect

DeviceUintSize

LayerPoint

LayerPoint4D

LayerRect

LayerSize

LayerToScrollTransform

LayerToWorldTransform

LayerTransform

LayoutPixel
Geometry in a stacking context's local coordinate space (logical pixels).

LayoutPoint

LayoutRect

LayoutSize

LayoutTransform

ScrollLayerPoint

ScrollLayerRect

ScrollLayerSize

ScrollToLayerTransform

ScrollToWorldTransform

TileOffset

TileSize

VRCompositorId

WebGLResult

WorldPoint

WorldPoint4D

WorldRect

WorldSize

WorldToLayerTransform