Crate webrender_traits

Modules

channel

Structs

AuxiliaryLists

AuxiliaryListsDescriptor
Describes the memory layout of the auxiliary lists.

BorderDisplayItem

BorderRadius

BorderSide

BoxShadowDisplayItem

BuiltDisplayList
A display list.

BuiltDisplayListDescriptor
Describes the memory layout of a display list.

ClipRegion

ColorF

ColorU

ComplexClipRegion

DevicePixel
Geometry in the coordinate system of the render target (screen or intermediate surface) in physical pixels.

DisplayItem

DisplayListBuilder

Epoch

ExternalEvent
An opaque pointer-sized value.

ExternalImageId
An arbitrary identifier for an external image provided by the application. It must be a unique identifier for each external image.

FontKey

GlyphDimensions

GlyphInstance

GlyphKey

GradientDisplayItem

GradientStop

IdNamespace

IframeDisplayItem

ImageDescriptor

ImageDisplayItem

ImageKey

ImageMask

ItemRange

LayerPixel
Geometry in a layer's local coordinate space (logical pixels).

NativeFontHandle
Native fonts are not used on Linux; all fonts are raw.

PipelineId

PushScrollLayerItem

PushStackingContextDisplayItem

RadialGradientDisplayItem

RectangleDisplayItem

RenderApi

RenderApiSender

ResourceId

ScrollLayerId

ScrollLayerPixel
Geometry in a layer's scrollable parent coordinate space (logical pixels).

ScrollLayerState

ServoScrollRootId

StackingContext

TextDisplayItem

WebGLBufferId

WebGLContextId

WebGLDisplayItem

WebGLFramebufferId

WebGLProgramId

WebGLRenderbufferId

WebGLShaderId

WebGLTextureId

WorldPixel
Geometry in the document's coordinate space (logical pixels).

YuvImageDisplayItem

Enums

ApiMsg

BorderStyle

BoxShadowClipMode

DisplayListMode

FilterOp

FontRenderMode

FragmentType

ImageData

ImageFormat

ImageRendering

MixBlendMode

RendererKind

ScrollEventPhase

ScrollLayerInfo

ScrollLocation

ScrollPolicy

SpecificDisplayItem

VRCompositorCommand

WebGLCommand

WebGLError

WebGLFramebufferBindingRequest

WebGLParameter

WebGLShaderParameter

YuvColorSpace

Traits

RenderDispatcher

RenderNotifier

VRCompositorHandler

Functions

as_scroll_parent_rect

device_length

Type Definitions

DeviceIntLength

DeviceIntPoint

DeviceIntRect

DeviceIntSize

DevicePoint

DeviceRect

DeviceSize

DeviceUintPoint

DeviceUintRect

DeviceUintSize

LayerPoint

LayerPoint4D

LayerRect

LayerSize

LayerToScrollTransform

LayerToWorldTransform

LayerTransform

LayoutPixel
Geometry in a stacking context's local coordinate space (logical pixels).

LayoutPoint

LayoutRect

LayoutSize

LayoutTransform

ScrollLayerPoint

ScrollLayerRect

ScrollLayerSize

ScrollToLayerTransform

ScrollToWorldTransform

VRCompositorId

WebGLResult

WorldPoint

WorldPoint4D

WorldRect

WorldSize

WorldToLayerTransform