Crate webrender_traits

Re-exports

pub use crate::rendering_context::RenderingContext;

Modules

display_list

Defines data structures which are consumed by the Compositor.

rendering_context

Structs

CompositorHitTestResult

The result of a hit test in the compositor.

CrossProcessCompositorApi

A mechanism to send messages from ScriptThread to the parent process’ WebRender instance.

UntrustedNodeAddress

The address of a node. Layout sends these back. They must be validated via from_untrusted_node_address before they can be used, because we do not trust layout.

WebrenderExternalImageHandlers

WebRender External Image Handler implementation.

WebrenderExternalImageRegistry

List of Webrender external images to be shared among all external image consumers (WebGL, Media, WebGPU). It ensures that external image identifiers are unique.

Enums

CrossProcessCompositorMessage

ImageUpdate

Serializable image updates that must be performed by WebRender.

SerializableImageData

Serialized ImageData. It contains IPC byte channel receiver to prevent from loading bytes too slow.

WebrenderImageHandlerType

Type of Webrender External Image Handler.

WebrenderImageSource

Traits

WebrenderExternalImageApi

This trait is used as a bridge between the different GL clients in Servo that handles WebRender ExternalImages and the WebRender ExternalImageHandler API. This trait is used to notify lock/unlock messages and get the required info that WR needs.