Enum webrender_traits::ApiMsg

pub enum ApiMsg {
    AddRawFont(FontKey, Vec<u8>),
    AddNativeFont(FontKey, NativeFontHandle),
    GetGlyphDimensions(Vec<GlyphKey>, IpcSender<Vec<Option<GlyphDimensions>>>),
    AddImage(ImageKey, u32, u32, Option<u32>, ImageFormat, Vec<u8>),
    UpdateImage(ImageKey, u32, u32, ImageFormat, Vec<u8>),
    DeleteImage(ImageKey),
    CloneApi(IpcSender<IdNamespace>),
    SetRootStackingContext(StackingContextId, ColorF, Epoch, PipelineId, Size2D<f32>, Vec<(StackingContextId, StackingContext)>, Vec<(DisplayListId, BuiltDisplayListDescriptor)>, AuxiliaryListsDescriptor),
    SetRootPipeline(PipelineId),
    Scroll(Point2D<f32>, Point2D<f32>, ScrollEventPhase),
    TickScrollingBounce,
    TranslatePointToLayerSpace(Point2D<f32>, IpcSender<(Point2D<f32>, PipelineId)>),
    GetScrollLayerState(IpcSender<Vec<ScrollLayerState>>),
    RequestWebGLContext(Size2D<i32>, GLContextAttributes, IpcSender<Result<(WebGLContextId, GLLimits), String>>),
    ResizeWebGLContext(WebGLContextId, Size2D<i32>),
    WebGLCommand(WebGLContextId, WebGLCommand),
}

Variants
AddRawFont(FontKey, Vec<u8>)AddNativeFont(FontKey, NativeFontHandle)GetGlyphDimensions(Vec<GlyphKey>, IpcSender<Vec<Option<GlyphDimensions>>>)

Gets the glyph dimensions

AddImage(ImageKey, u32, u32, Option<u32>, ImageFormat, Vec<u8>)

Adds an image from the resource cache.

UpdateImage(ImageKey, u32, u32, ImageFormat, Vec<u8>)

Updates the the resource cache with the new image data.

DeleteImage(ImageKey)

Drops an image from the resource cache.

CloneApi(IpcSender<IdNamespace>)SetRootStackingContext(StackingContextId, ColorF, Epoch, PipelineId, Size2D<f32>, Vec<(StackingContextId, StackingContext)>, Vec<(DisplayListId, BuiltDisplayListDescriptor)>, AuxiliaryListsDescriptor)

Supplies a new frame to WebRender.

The first StackingContextId describes the root stacking context. The actual stacking contexts are supplied as the sixth parameter, while the display lists that make up those stacking contexts are supplied as the seventh parameter.

After receiving this message, WebRender will read the display lists, followed by the auxiliary lists, from the payload channel.

SetRootPipeline(PipelineId)Scroll(Point2D<f32>, Point2D<f32>, ScrollEventPhase)TickScrollingBounceTranslatePointToLayerSpace(Point2D<f32>, IpcSender<(Point2D<f32>, PipelineId)>)GetScrollLayerState(IpcSender<Vec<ScrollLayerState>>)RequestWebGLContext(Size2D<i32>, GLContextAttributes, IpcSender<Result<(WebGLContextId, GLLimits), String>>)ResizeWebGLContext(WebGLContextId, Size2D<i32>)WebGLCommand(WebGLContextId, WebGLCommand)