Enum webrender_traits::ApiMsg

pub enum ApiMsg {
    AddRawFont(FontKey, Vec<u8>),
    AddNativeFont(FontKey, NativeFontHandle),
    GetGlyphDimensions(Vec<GlyphKey>, MsgSender<Vec<Option<GlyphDimensions>>>),
    AddImage(ImageKey, ImageDescriptor, ImageData),
    UpdateImage(ImageKey, ImageDescriptor, Vec<u8>),
    DeleteImage(ImageKey),
    CloneApi(MsgSender<IdNamespace>),
    SetRootDisplayList(Option<ColorF>, Epoch, PipelineId, LayoutSize, BuiltDisplayListDescriptor, AuxiliaryListsDescriptor),
    SetRootPipeline(PipelineId),
    Scroll(ScrollLocation, WorldPoint, ScrollEventPhase),
    ScrollLayersWithScrollId(LayoutPoint, PipelineId, ServoScrollRootId),
    TickScrollingBounce,
    TranslatePointToLayerSpace(WorldPoint, MsgSender<(LayoutPoint, PipelineId)>),
    GetScrollLayerState(MsgSender<Vec<ScrollLayerState>>),
    RequestWebGLContext(DeviceIntSize, GLContextAttributes, MsgSender<Result<(WebGLContextId, GLLimits), String>>),
    ResizeWebGLContext(WebGLContextId, DeviceIntSize),
    WebGLCommand(WebGLContextId, WebGLCommand),
    GenerateFrame,
    VRCompositorCommand(WebGLContextId, VRCompositorCommand),
    ExternalEvent(ExternalEvent),
    ShutDown,
}

Variants
AddRawFont(FontKey, Vec<u8>)AddNativeFont(FontKey, NativeFontHandle)GetGlyphDimensions(Vec<GlyphKey>, MsgSender<Vec<Option<GlyphDimensions>>>)

Gets the glyph dimensions

AddImage(ImageKey, ImageDescriptor, ImageData)

Adds an image from the resource cache.

UpdateImage(ImageKey, ImageDescriptor, Vec<u8>)

Updates the the resource cache with the new image data.

DeleteImage(ImageKey)

Drops an image from the resource cache.

CloneApi(MsgSender<IdNamespace>)SetRootDisplayList(Option<ColorF>, Epoch, PipelineId, LayoutSize, BuiltDisplayListDescriptor, AuxiliaryListsDescriptor)

Supplies a new frame to WebRender.

After receiving this message, WebRender will read the display list, followed by the auxiliary lists, from the payload channel.

SetRootPipeline(PipelineId)Scroll(ScrollLocation, WorldPoint, ScrollEventPhase)ScrollLayersWithScrollId(LayoutPoint, PipelineId, ServoScrollRootId)TickScrollingBounceTranslatePointToLayerSpace(WorldPoint, MsgSender<(LayoutPoint, PipelineId)>)GetScrollLayerState(MsgSender<Vec<ScrollLayerState>>)RequestWebGLContext(DeviceIntSize, GLContextAttributes, MsgSender<Result<(WebGLContextId, GLLimits), String>>)ResizeWebGLContext(WebGLContextId, DeviceIntSize)WebGLCommand(WebGLContextId, WebGLCommand)GenerateFrameVRCompositorCommand(WebGLContextId, VRCompositorCommand)ExternalEvent(ExternalEvent)

An opaque handle that must be passed to the render notifier. It is used by Gecko to forward gecko-specific messages to the render thread preserving the ordering within the other messages.

ShutDown