Struct webrender_traits::RenderApi

pub struct RenderApi {
    pub api_sender: MsgSender<ApiMsg>,
    pub payload_sender: PayloadSender,
    pub id_namespace: IdNamespace,
    pub next_id: Cell<ResourceId>,
}

Fields
api_sender: MsgSender<ApiMsg> payload_sender: PayloadSender id_namespace: IdNamespace next_id: Cell<ResourceId>
Methods

impl RenderApi

fn clone_sender(&self) -> RenderApiSender

fn add_raw_font(&self, bytes: Vec<u8>) -> FontKey

fn add_native_font(&self, native_font_handle: NativeFontHandle) -> FontKey

fn get_glyph_dimensions(&self,
glyph_keys: Vec<GlyphKey>)
-> Vec<Option<GlyphDimensions>>

Gets the dimensions for the supplied glyph keys

Note: Internally, the internal texture cache doesn't store 'empty' textures (height or width = 0) This means that glyph dimensions e.g. for spaces (' ') will mostly be None.

fn alloc_image(&self) -> ImageKey

Creates an ImageKey.

fn add_image(&self, descriptor: ImageDescriptor, data: ImageData) -> ImageKey

Adds an image and returns the corresponding ImageKey.

fn update_image(&self,
key: ImageKey,
descriptor: ImageDescriptor,
bytes: Vec<u8>)

Updates a specific image.

Currently doesn't support changing dimensions or format by updating.

fn delete_image(&self, key: ImageKey)

Deletes the specific image.

fn set_root_pipeline(&self, pipeline_id: PipelineId)

Sets the root pipeline.

Examples

let (mut renderer, sender) = webrender::renderer::Renderer::new(opts);
let api = sender.create_api();
...
let pipeline_id = PipelineId(0,0);
api.set_root_pipeline(pipeline_id);

fn set_root_display_list(&self,
background_color: Option<ColorF>,
epoch: Epoch,
viewport_size: LayoutSize,
builder: DisplayListBuilder)

Supplies a new frame to WebRender.

Non-blocking, it notifies a worker process which processes the display list. When it's done and a RenderNotifier has been set in webrender::renderer::Renderer, new_frame_ready() gets called.

Note: Scrolling doesn't require an own Frame.

Arguments:

background_color: The background color of this pipeline.

epoch: The unique Frame ID, monotonically increasing.

pipeline_id: The ID of the pipeline that is supplying this display list.

viewport_size: The size of the viewport for this frame.

display_list: The root Display list used in this frame.

auxiliary_lists: Various items that the display lists and stacking contexts reference.

fn scroll(&self,
scroll_location: ScrollLocation,
cursor: WorldPoint,
phase: ScrollEventPhase)

Scrolls the scrolling layer under the cursor

Webrender looks for the layer closest to the user which has ScrollPolicy::Scrollable set.

fn scroll_layers_with_scroll_root_id(&self,
new_scroll_origin: LayoutPoint,
pipeline_id: PipelineId,
scroll_root_id: ServoScrollRootId)

fn tick_scrolling_bounce_animations(&self)

fn translate_point_to_layer_space(&self,
point: &WorldPoint)
-> (LayoutPoint, PipelineId)

Translates a point from viewport coordinates to layer space

fn get_scroll_layer_state(&self) -> Vec<ScrollLayerState>

fn request_webgl_context(&self,
size: &DeviceIntSize,
attributes: GLContextAttributes)
-> Result<(WebGLContextId, GLLimits), String>

fn resize_webgl_context(&self, context_id: WebGLContextId, size: &DeviceIntSize)

fn send_webgl_command(&self, context_id: WebGLContextId, command: WebGLCommand)

fn generate_frame(&self)

fn send_vr_compositor_command(&self,
context_id: WebGLContextId,
command: VRCompositorCommand)

fn send_external_event(&self, evt: ExternalEvent)

fn shut_down(&self)